+++ android/src/main/java/com/llfbandit/record/MethodCallHandlerImpl.java
@@ -1,28 +1,28 @@
 package com.llfbandit.record;
 
 import android.Manifest;
+import android.app.Activity;
 import android.content.Context;
 import android.content.pm.PackageManager;
-import android.media.MediaRecorder;
-import android.os.Environment;
-import android.util.Log;
+
 import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import java.io.IOException;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.HashMap;
+import io.flutter.plugin.common.PluginRegistry;
 
-public class MethodCallHandlerImpl implements MethodCallHandler {
+public class MethodCallHandlerImpl implements MethodCallHandler, PluginRegistry.RequestPermissionsResultListener {
   private static final String LOG_TAG = "Record";
-  private Recorder recorder = new Recorder();
+  private static final int RECORD_AUDIO_REQUEST_CODE = MethodCallHandlerImpl.class.hashCode() + 43;
 
-  private final Context context;
+  private final Activity activity;
+  private Result pendingPermResult;
+  private final Recorder recorder = new Recorder();
 
-  MethodCallHandlerImpl(Context context) {
-    this.context = context;
+  MethodCallHandlerImpl(Activity activity) {
+    this.activity = activity;
   }
 
   void close() {
@@ -35,7 +35,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
       case "start":
         recorder.start(
           (String) call.argument("path"),
-          (int) call.argument("outputFormat"),
           (int) call.argument("encoder"),
           (int) call.argument("bitRate"),
           (double) call.argument("samplingRate"),
@@ -56,17 +55,37 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
     }
   }
 
+  @Override
+  public boolean onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+    if (requestCode == RECORD_AUDIO_REQUEST_CODE) {
+      if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+        pendingPermResult.success(true);
+      } else {
+        pendingPermResult.error("-2", "Permission denied", null);          
+      }
+      pendingPermResult = null;
+      return true;
+    }
+
+    return false;
+  }
+
   private void hasPermission(@NonNull Result result) {
-    final PackageManager pm = context.getPackageManager();
-    final String pkgName = context.getPackageName();
-
-    int perm = pm.checkPermission(Manifest.permission.RECORD_AUDIO, pkgName);
-    boolean permGranted = perm == PackageManager.PERMISSION_GRANTED;
-
-    if (permGranted) {
-      result.success(true);
+    if (!isPermissionGranted(Manifest.permission.RECORD_AUDIO)) {
+      pendingPermResult = result;
+      askForPermission(Manifest.permission.RECORD_AUDIO, RECORD_AUDIO_REQUEST_CODE);
+      return;
     } else {
-      result.error("-2", "Permission denied", null);
-    }    
+      result.success(true);
+    }
+  }
+
+  private boolean isPermissionGranted(String permissionName) {
+    int result = ActivityCompat.checkSelfPermission(activity, permissionName);
+    return result == PackageManager.PERMISSION_GRANTED;
+  }
+
+  private void askForPermission(String permissionName, int requestCode) {
+    ActivityCompat.requestPermissions(activity, new String[]{permissionName}, requestCode);
   }
 }