+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0
+
+* Flutter 1.12.0 is now minimum version (Flutter plugin API v2).
+* Finalize Record API.
+* Request permission on Android too.
+* Automatically stop recording when leaving activity / app.
+* Add all codecs with cross platform compatibility.
+
 ## 0.2.1
 
 * Fix broken build on Android.